fix(ci): unnecessary_cast on ecount() — clippy 1.95 hotfix
Rust 1.95's clippy::unnecessary_cast flags `g.ecount() as usize`
because `Graph::ecount()` already returns usize. Strip the cast at the
three sites (one bench, one property test, one biconnected doctest).

+++ benches/bench_hits_weighted.rs
@@ -29,7 +29,7 @@ fn bench_weighted_hits_karate(c: &mut Criterion) {
     // Undirected karate under unit weights → exercises the shifted
     // power-iter path on (W + I).
     let g = karate();
-    let weights = vec![1.0; g.ecount() as usize];
+    let weights = vec![1.0; g.ecount()];
     c.bench_function(
         "hits_weighted/karate (34v 78e, undirected, unit weights)",
         |b| {
